// src/app/models/Customers.ts
export class Customers{
    customer_id?: number;
    customer_id_external?: number;
    client_id?: number;
    customer_name?: string;
    customer_user_name?: string;
    customer_cnic?: string;
    cusotmer_mobilenum?: string;
    customer_address?: string;
    package_id?: number;
    customer_due_day?: number;
    join_date?: string;
    customer_status?: string;
    selected: boolean = false;
    package_name?: string;
}